move cache_transform to its own file
